requirements.txt
streamlit
pandas
numpy
scikit-learn
supabase
python-dotenv
reportlab
groq
cohere
google-generativeai
joblib
